[MachO] Filling in some structs I missed in r269499
I missed the fvmlib_command and the sub_framework_command, as well as a few uses of the dylib_command, dylinker_command, and linkedit_data_command.

This should now be a pretty complete listing. The only case I'm not sure about is LC_PREPAGE which doesn't seem to be referenced directly anywhere in LLVM.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@269513 91177308-0d34-0410-b5e6-96231b3b80d8